SVN Commit: Bug#257426 - Get LRGs working on Derby
https://bugs.eclipse.org/bugs/show_bug.cgi?id=257426

 

Tested our JPA and core LRGs on Derby and DB2.  Got most of the tests passing.  There are still a couple of failures, but much fewer than when started.  The remaining failures are from a update-all temp table issue in EclipseLink, and some bugs in Derby.

 

I added the platform result to our wiki test status page.  This would be a good place for people to add test results to platforms or tests not tested by the nightly test runs.

 

http://wiki.eclipse.org/EclipseLink/Development/Testing/Status

 

Changes:

- Auto default UpdateAll temp table type to String if unknown.

- Switch to OraclePlatform in auto config if cannot load Oracle10Platform.

- Removed incorrect toUppercase() in order by of embeddable

- Fixed Derby platform field type defaults and toNumber function.

 

- Added Derby option to test browser.

- Consolidated testing checks for select for update and stored procedure

support.

- Fixed sleep issue in legacy model.

- Fixed reserved word SECOND and BOOLEAN in Sequence, NoIdentity, Types model.

- Removed hint from custom SQL model.

- Removed comment from preventing DB2Function test from running on DB2.

- Fixed supported platforms in various tests.

- Fixed bad SQL in QueryOptions test.

- Fixed Naboo city.

 

- Fixed constraint name sizes in JPA table creators.

- Do not created Oracle stored procedures when not on Oracle (or MySQL...).

- Fixed XML address to mark the type as transient on the get method, not field.

- Some cleanup.

- Changed system millis id to int to avoid database truncation.

+++ jpa/eclipselink.jpa.test/src/org/eclipse/persistence/testing/tests/jpa/ddlgeneration/DDLGenerationJUnitTestSuite.java
@@ -88,7 +88,7 @@ public class DDLGenerationJUnitTestSuite extends JUnitTestCase {
         beginTransaction(em);
         try {
             CKeyEntityA a = new CKeyEntityA("Wonseok", "Kim");
-            long seq = System.currentTimeMillis(); // just to get unique value :-)
+            int seq = (int)System.currentTimeMillis(); // just to get unique value :-)
             CKeyEntityB b = new CKeyEntityB(new CKeyEntityBPK(seq, "B1209"));
             //set unique keys
             b.setUnq1("u0001");
@@ -345,7 +345,7 @@ public class DDLGenerationJUnitTestSuite extends JUnitTestCase {
         beginTransaction(em);
         // let's see if a subclass entity is persisted and found well
         try {
-            long seq = System.currentTimeMillis(); // just to get unique value :-)
+            int seq = (int)System.currentTimeMillis(); // just to get unique value :-)
             String code = "B1215";
             CKeyEntityB2 b = new CKeyEntityB2(new CKeyEntityBPK(seq, code));
             //set unique keys
